+++ src/main/java/net/jasper/mod/mixins/accessors/ScreenAccessor.java
@@ -1,6 +1,5 @@
 package net.jasper.mod.mixins.accessors;
 
-import com.google.common.collect.Lists;
 import net.minecraft.client.gui.Drawable;
 import net.minecraft.client.gui.screen.Screen;
 import org.spongepowered.asm.mixin.Mixin;